Refactor assets files
