+++ public/images/discover-furniture.jpg
